Move ISmsNotification to namespace Digipost.Api.Client.Send
SmsNotification is only used in conjuction with sending messages to
Digipost and thus should live in the Send-module.

This is minor breaking change where the using namespace need to be
updated.

I also did some minor code enhances..
