+++ Digipost.Api.Client.Tests/Smoke/ClientSmokeTests.cs
@@ -9,20 +9,20 @@ namespace Digipost.Api.Client.Tests.Smoke
 {
     public class ClientSmokeTests
     {
-        private static ClientSmokeTestHelper client;
-        private static ClientSmokeTestHelper clientWithoutDataTypes;
-        
+        private static ClientSmokeTestHelper _client;
+        private static ClientSmokeTestHelper _clientWithoutDataTypes;
+
         public ClientSmokeTests()
         {
             var sender = SenderUtility.GetSender(TestEnvironment.Qa);
-            client = new ClientSmokeTestHelper(sender);
-            clientWithoutDataTypes = new ClientSmokeTestHelper(sender, true);
+            _client = new ClientSmokeTestHelper(sender);
+            _clientWithoutDataTypes = new ClientSmokeTestHelper(sender, true);
         }
         
         [Fact(Skip = "SmokeTest")]
         public void Can_identify_user()
         {
-            client
+            _client
                 .Create_identification_request()
                 .SendIdentification()
                 .Expect_identification_to_have_status(IdentificationResultType.DigipostAddress);
@@ -31,7 +31,7 @@ namespace Digipost.Api.Client.Tests.Smoke
         [Fact(Skip = "SmokeTest")]
         public void Can_Search()
         {
-            client
+            _client
                 .Create_search_request()
                 .Expect_search_to_have_result();
         }
@@ -39,33 +39,33 @@ namespace Digipost.Api.Client.Tests.Smoke
         [Fact(Skip = "SmokeTest")]
         public void Can_send_document_digipost_user()
         {
-            client
+            _client
                 .Create_message_with_primary_document()
                 .To_Digital_Recipient()
                 .SendMessage()
                 .Expect_message_to_have_status(MessageStatus.Delivered);
         }
-        
+
         [Fact(Skip = "SmokeTest")]
         public void Can_send_document_with_raw_datatype_to_digipost_user()
         {
-            var raw = "<?xml version=\"1.0\" encoding=\"utf-8\"?><externalLink xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" xmlns:xsd=\"http://www.w3.org/2001/XMLSchema\" xmlns=\"http://api.digipost.no/schema/datatypes\"><url>https://www.test.no</url><description>This was raw string</description></externalLink>";
-            
-            clientWithoutDataTypes
+            const string raw = "<?xml version=\"1.0\" encoding=\"utf-8\"?><externalLink xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" xmlns:xsd=\"http://www.w3.org/2001/XMLSchema\" xmlns=\"http://api.digipost.no/schema/datatypes\"><url>https://www.test.no</url><description>This was raw string</description></externalLink>";
+
+            _clientWithoutDataTypes
                 .CreateMessageWithPrimaryDataTypeDocument(raw)
                 .To_Digital_Recipient()
                 .SendMessage()
                 .Expect_message_to_have_status(MessageStatus.Delivered);
         }
-        
+
         [Fact(Skip = "SmokeTest")]
         public void Can_send_document_with_object_datatype_to_digipost_user()
         {
-            
-            ExternalLink externalLink = new ExternalLink {Url = "https://www.test.no", Description = "This is a link"};
-            string linkXml = SerializeUtil.Serialize(externalLink);
 
-            client 
+            var externalLink = new ExternalLink {Url = "https://www.test.no", Description = "This is a link"};
+            var linkXml = SerializeUtil.Serialize(externalLink);
+
+            _client
                 .CreateMessageWithPrimaryDataTypeDocument(linkXml)
                 .To_Digital_Recipient()
                 .SendMessage()
